% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmeterbarplotter.r
\name{jbp_aggr}
\alias{jbp_aggr}
\title{aggregate dataframe of csv files.}
\usage{
jbp_aggr(data = data.frame())
}
\arguments{
\item{data}{dataframe returned from jbp_read}
}
\value{
DescTools MeanCI.
}
\description{
aggregate dataframe of csv files.
}
